Refactored to improve project structure and separate styling.
