+++ src/App.css
@@ -1,98 +1,94 @@
 .App {
-  text-align: center;
+    text-align: center;
 }
 
 .App-logo {
-  height: 40vmin;
-  pointer-events: none;
+    height: 40vmin;
+    pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
+    .App-logo {
+        animation: App-logo-spin infinite 20s linear;
+    }
 }
 
 .App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
+    background-color: #282c34;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
 }
 
-.App-link {
-  color: #61dafb;
+body {
+    font-family: Arial, sans-serif;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.page-tile {
+    color: #FF2625;
+    font-weight: 700;
+    text-transform: capitalize;
+    font-size: 50px;
+    margin-bottom: 15px;
+    margin-top: 0;
 }
 
-/*Image styling*/
-.banner-img {
-  width: 100%;
-  height: 405px;
-  object-fit: fill;
-  border-radius: 10px;
+p {
+    max-width: 80vw;
+    line-height: 1.6;
+    font-size: 20px;
 }
 
-.detail-image {
-  width: 729px;
-  height: 742px;
+
+/* Common Flex Grid Styles */
+.flex-grid {
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    gap: 15px;
 }
 
-/*Home*/
-.exercise-btn {
-  margin-top: 25px !important;
-  text-decoration: none !important;
-  width: 200px;
-  text-align: center;
-  color: white !important;
-  font-size: 18px !important;
-  text-transform: none !important;
-  border-radius: 4px !important;
+@media (min-width: 992px) {
+    .flex-grid {
+        flex-direction: row !important;
+    }
 }
 
 /*Exercise buttons*/
 .pill {
-  padding: 7px 10px !important;
-  margin-top: 10px !important;
-  margin-right: 18px !important;
-  color: #fff !important;
-  font-size: 14px !important;
-  border-radius: 20px !important;
-  text-transform: uppercase !important;
+    padding: 7px 10px !important;
+    margin-top: 10px !important;
+    margin-right: 18px !important;
+    color: #fff !important;
+    font-size: 14px !important;
+    border-radius: 20px !important;
+    text-transform: uppercase !important;
 }
 
 .body-part-btn {
-  background-color: #0000FF !important;
+    background-color: #0000FF !important;
 }
 
 .body-part-btn:hover {
-  background-color: #0000FF !important;
+    background-color: #0000FF !important;
 }
 
 .target-muscle-btn {
-  background-color: #008000 !important;
+    background-color: #008000 !important;
 }
 
 .target-muscle-btn:hover {
-  background-color: #008000 !important;
+    background-color: #008000 !important;
 }
 
 .equipment-btn {
-  background-color: #A52A2A !important;
+    background-color: #A52A2A !important;
 }
 
 .equipment-btn:hover {
-  background-color: #A52A2A !important;
+    background-color: #A52A2A !important;
 }
