+++ src/components/Exercise/ExerciseCard/ExerciseCard.css
@@ -1,0 +1,56 @@
+.exercise-card-link {
+    text-decoration: none;
+}
+
+.exercise-card {
+    border: 1px solid #ccc;
+    border-radius: 8px;
+    display: flex;
+    align-items: center;
+    text-decoration: none;
+    height: 100%;
+    overflow: hidden;
+    flex: 0 450px;
+    width: 29vw;
+}
+
+@media (max-width: 992px) {
+    .exercise-card {
+        width: 80vw;
+    }
+}
+
+.exercise-img {
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+    margin: 10px 15px;
+}
+
+@media (min-width: 992px) {
+    .exercise-img {
+        width: 150px;
+        height: 150px;
+    }
+}
+
+.card-info-section {
+    margin: 20px 20px 10px 20px;
+    overflow: hidden;
+    flex: 1;
+    flex-direction: column;
+}
+
+.exercise-name {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    display: block;
+    max-width: 100%;
+    color: #000;
+    font-weight: bold;
+    font-size: 20px;
+    margin-top: 0;
+    margin-bottom: 10px;
+    text-transform: capitalize;
+}
