+++ src/components/Exercise/ExerciseList/ExerciseList.css
@@ -1,0 +1,23 @@
+/* Specific Styles for Exercise Flex Grid */
+.exercises-flex-grid {
+    padding-right: 30px;
+    padding-left: 30px;
+    justify-content: space-around;
+    align-content: space-around;
+}
+
+@media (min-width: 992px) {
+    .exercises-flex-grid {
+        gap: 20px;
+    }
+}
+
+.no-exercise-found {
+    font-size: 25px;
+    font-weight: 500;
+    text-transform: capitalize;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 46px;
+}
