+++ src/components/Exercise/Filter/ExerciseFilter.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import {Stack} from '@mui/material';
+import BodyPartFilter from "./BodyPartFilter/BodyPartFilter";
+import SearchExercise from "./SearchExercise/SearchExercise";
+
+const ExerciseFilter = () => {
+    return (
+        <Stack alignItems="center" justifyContent="center">
+            <SearchExercise/>
+            <BodyPartFilter/>
+        </Stack>
+    );
+};
+
+export default ExerciseFilter;
