+++ src/components/Exercise/Filter/SearchExercise/SearchExercise.css
@@ -1,0 +1,34 @@
+.searchTextField {
+    height: 76px;
+    width: 80vw;
+    background-color: #fff;
+    border-radius: 40px;
+}
+
+.searchInput {
+    font-weight: 700 !important;
+    border: none !important;
+    border-radius: 4px !important;
+}
+
+.search-btn {
+    background-color: blue !important;
+    color: white !important;
+    text-transform: none !important;
+    height: 56px !important;
+    width: 173px !important;
+    position: absolute !important;
+    right: 0 !important;
+    font-size: 20px !important;
+}
+
+.search-btn:hover {
+    background-color: dodgerblue !important;
+}
+
+@media (max-width: 599px) {
+    .search-btn {
+        width: 90px !important;
+        font-size: 14px !important;
+    }
+}
